Add splitting theorem
In order to prove the normality of the extension in reverse.lean,
introduce theorem 16.3.2 from Artin (Splitting Theorem).

Currently, there are three sorries:
(1) I don't know how to make sense of "adjoining all the roots of
a polynomial f" in lean.
(2) notion (1) is necessary for proving that F is a splitting
field of it's fixed field
(3) notion (1) is necessary for proving the splitting lemma

Nonetheless, I think I have all the statements there.